pyproject.toml
[project]
name = "device-gateway"
version = "1.0.0"
description = "Device Gateway"
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.70.0",
    "grpcio-reflection>=1.70.0",
    "protobuf>=5.29.3",
    "ruff>=0.9.4",
    "rustworkx>=0.16.0",
    "qiskit~=1.3.2",
    "qulacs>=0.6.11",
    "networkx>=3.4.2",
    "matplotlib>=3.8.0",
    "qiskit-qasm3-import>=0.5.1",
    "pyyaml>=6.0.2",
    "python-json-logger>=2.0.7",
]

[dependency-groups]
test = ["pytest", "pytest-cov", "mypy"]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/device_gateway"]

[tool.mypy]
ignore_missing_imports = true


[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=xml"
pythonpath = ["src","src/device_gateway","src/device_gateway/qpu_interface","src/device_gateway/backend","src/device_gateway/circuit"]
